gotssite/requirements.txt
Django
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
drf-yasg==1.17.1
plotly==4.12.0
numpy
matplotlib
reportlab